Pitch, Roll, Yaw as opposed to X,Y,Z for rotation
After going through the Limelight documentation, I came across "botpose" and noticed it states it returns Rotation in terms of "X, Y, Z". This was perplexing as it is very different from the expected "pitch, roll, yaw" and has caused me some confusion. After realizing this must be a minor mistake, the error vexed me for a while causing difficulty concentrating on the rest of the file. Thank you for your time and amazing work on this documentation as a whole: it has been extremely helpful!